pyproject.toml
[project]
name = "simple_rag"
version = "0.0.1"
requires-python = ">=3.10"

dependencies = [
    "langchain",
    "langchain-openai",
    "langchain_community",
    "langchainhub",
    "langgraph",
    "chromadb",
    "tiktoken",
    "tavily-python",
    "youtube-transcript-api",
    "python-dotenv",
    "bs4"
]

[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.autopep8]
max_line_length = 120

[tool.black]
line-length = 120
list-line-length = 1